Commands to run DKAN tests for dkan2.
